+++ packages/stream_video_flutter/ios/Classes/StreamVideoFlutterPlugin.swift
@@ -1,0 +1,19 @@
+import Flutter
+import UIKit
+
+public class StreamVideoFlutterPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "stream_video_flutter", binaryMessenger: registrar.messenger())
+    let instance = StreamVideoFlutterPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    switch call.method {
+    case "getPlatformVersion":
+      result("iOS " + UIDevice.current.systemVersion)
+    default:
+      result(FlutterMethodNotImplemented)
+    }
+  }
+}