+++ packages/stream_video_flutter/lib/src/call_background/model/service_type.dart
@@ -1,0 +1,4 @@
+enum ServiceType {
+  call,
+  screenSharing,
+}