+++ packages/stream_video_flutter/lib/src/call_participants/regular_call_participants_content.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import '../../stream_video_flutter.dart';
 import 'layout/call_participants_grid_view.dart';
 import 'layout/call_participants_spotlight_view.dart';
-import 'layout/participant_layout_mode.dart';
 
 /// A widget that represents the main area of the call when nobody is
 /// sharing their screen.